app/requirements.txt
Flask>=2.1.0
Pillow==9.5.0
torch>=2.0.1
torchvision>=0.16.0
numpy<2
gunicorn==20.1.0